+++ web/src/test/java/org/springframework/security/web/aot/hint/PublicKeyCredentialUserEntityRuntimeHintsTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2025 the original author or authors.
+ * Copyright 2002-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.security.web.webauthn.aot;
+package org.springframework.security.web.aot.hint;
 
 import java.util.stream.Stream;
 
@@ -31,11 +31,11 @@ import org.springframework.util.ClassUtils;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
- * Tests for {@link UserCredentialRuntimeHints}
+ * Tests for {@link PublicKeyCredentialUserEntityRuntimeHints}
  *
  * @author Max Batischev
  */
-public class UserCredentialRuntimeHintsTests {
+public class PublicKeyCredentialUserEntityRuntimeHintsTests {
 
 	private final RuntimeHints hints = new RuntimeHints();
 
@@ -47,14 +47,13 @@ public class UserCredentialRuntimeHintsTests {
 	}
 
 	@ParameterizedTest
-	@MethodSource("getClientRecordsSqlFiles")
-	void credentialRecordsSqlFilesHasHints(String schemaFile) {
+	@MethodSource("getUserEntitiesSqlFiles")
+	void userEntitiesSqlFilesHasHints(String schemaFile) {
 		assertThat(RuntimeHintsPredicates.resource().forResource(schemaFile)).accepts(this.hints);
 	}
 
-	private static Stream<String> getClientRecordsSqlFiles() {
-		return Stream.of("org/springframework/security/user-credentials-schema.sql",
-				"org/springframework/security/user-credentials-schema-postgres.sql");
+	private static Stream<String> getUserEntitiesSqlFiles() {
+		return Stream.of("org/springframework/security/user-entities-schema.sql");
 	}
 
 }
