--- webauthn/src/main/resources/META-INF/spring/aot.factories
@@ -1,3 +1,0 @@
-org.springframework.aot.hint.RuntimeHintsRegistrar=\
-org.springframework.security.web.webauthn.aot.UserCredentialRuntimeHints,\
-org.springframework.security.web.webauthn.aot.PublicKeyCredentialUserEntityRuntimeHints
